Relocate 'omise-wc-gateway.php' to 'includes/gateway/class-omise-payment-creditcard.php' and do code refactoring.
Here, new payment gateway will be localted inside 'includes/gateway' and extend with 'Omise_Payment' base class.
