+++ omise-woocommerce.php
@@ -22,6 +22,7 @@ defined( 'OMISE_WOOCOMMERCE_TEXT_DOMAIN' ) || define( 'OMISE_WOOCOMMERCE_TEXT_DO
 
 require_once dirname( __FILE__ ) . '/includes/libraries/omise-php/lib/Omise.php';
 require_once dirname( __FILE__ ) . '/includes/libraries/omise-plugin/Omise.php';
+require_once dirname( __FILE__ ) . '/includes/gateway/class-omise-payment-creditcard.php';
 require_once dirname( __FILE__ ) . '/includes/classes/class-omise-charge.php';
 require_once dirname( __FILE__ ) . '/includes/classes/class-omise-hooks.php';
 require_once dirname( __FILE__ ) . '/includes/classes/class-omise-transfer.php';
@@ -29,12 +30,11 @@ require_once dirname( __FILE__ ) . '/includes/classes/class-omise-card-image.php
 
 require_once 'omise-util.php';
 require_once 'omise-api-wrapper.php';
-require_once 'omise-wc-gateway.php';
 require_once 'omise-wc-myaccount.php';
 require_once 'omise-wp-admin.php';
 
 add_action( 'init', 'register_omise_wc_gateway_post_type' );
-add_action( 'plugins_loaded', 'register_omise_wc_gateway_plugin', 0 );
+add_action( 'plugins_loaded', 'register_omise_creditcard', 0 );
 add_action( 'plugins_loaded', 'prepare_omise_myaccount_panel', 0 );
 
 // Include these files only when we are in the admin pages
